Division into definitions and other files
